+++ CHANGELOG.md
@@ -28,9 +28,9 @@ Unreleased and unplanned todos
 
 ## v0.3.9 (Unreleased)
 
-- [ ] Appname can be configured in Freyfile, defaults to git dir. Git dir is in init.paths
 - [ ] Remove invalid underscore prefix from a few public `Shell` methods
 - [ ] For connecting to all hosts, we'd need to duplicate STDIN. Alternatively: Make interactive host selection a list, vs checkbox, and remove 'all'
+- [x] Appname can be configured in Freyfile, defaults to git dir. Git dir is in init.paths
 - [x] Show endpoint if such output is available
 
 ## v0.3.8 (2016-03-18)